move DecisionTree's main method to examples
